Fundamental fix: Train additional LLM in float32 with AMP
Root cause: Training float16 models causes GradScaler unscale errors.

Changes:
- smollm2_cascade.py: Create additional LLM in float32 (not .half())
- cascade_trainer.py: Auto-enable AMP only for float32 models
- CLAUDE.md: Document dtype design principles for training

Design principle:
- Frozen base model: float16 (inference only, memory efficient)
- Trainable model: float32 (gradient stability, AMP compatible)

+++ experiments/smollm2_cascade.py
@@ -192,10 +192,12 @@ def main() -> None:
     print(f"ベースLLM: dim={llm_base.dim}, layers={llm_base.num_layers}, frozen=True")
 
     # 後付けLLMの作成（ベースモデルと同じアーキテクチャ）
+    # 注: 訓練対象モデルはfloat32で作成（勾配計算の安定性のため）
+    # AMPを使用する場合、TrainerがAMP管理下でfloat16計算を自動実行
     additional_model = create_llm_from_base(base_model, num_layers=args.additional_layers)
 
     if device == "cuda":
-        additional_model = additional_model.to(device).half()  # type: ignore[arg-type]
+        additional_model = additional_model.to(device)  # float32のまま
 
     llm_additional = LLM(additional_model)
